Uninstall `swr` + minor renaming and cleanup
