+++ next/components/molecules/sections/general/ArticlesListSection/BlogPostsList.tsx
@@ -12,7 +12,7 @@ import {
   blogPostsDefaultFilters,
   blogPostsFetcher,
   getBlogPostsQueryKey,
-} from '@/services/meili/fetchers/blogPostsFetcherReactQuery'
+} from '@/services/meili/fetchers/blogPostsFetcher'
 import { getCategoryColorLocalStyle } from '@/utils/colors'
 import { generateImageSizes } from '@/utils/generateImageSizes'
 import { getNumericLocalDate } from '@/utils/local-date'
