+++ next/package.json
@@ -65,7 +65,6 @@
     "rss": "^1.2.2",
     "smooth-scroll-into-view-if-needed": "^1.1.33",
     "string-sanitizer": "^2.0.2",
-    "swr": "^1.3.0",
     "tailwind-merge": "^1.11.0",
     "tailwind-scrollbar-hide": "^1.1.7",
     "tailwindcss-animate": "^1.0.7",
