--- next/utils/usePrevious.ts
@@ -1,9 +1,0 @@
-import { useEffect, useRef } from 'react'
-
-export default function usePrevious<T>(value: T): T | undefined {
-  const ref = useRef<T>()
-  useEffect(() => {
-    ref.current = value
-  })
-  return ref.current
-}
